♻️ #85 refactor: GetTodos APi 관련  타입정리리
