+++ lib/utils/findMatchingTodoQueries.ts
@@ -1,12 +1,12 @@
 import { InfiniteData, QueryKey, useQueryClient } from '@tanstack/react-query';
-import { TodosResponse } from '../hooks/useUpdateTodoMutation';
+import { GetTodosResponse } from '../types/todos';
 
 export const isTodosQueryKey = (queryKey: QueryKey): queryKey is ['todos'] | ['todos', Record<string, unknown>] => {
   return Array.isArray(queryKey) && queryKey[0] === 'todos';
 };
 
 export const findMatchingTodoQueries = (queryClient: ReturnType<typeof useQueryClient>) => {
-  return queryClient.getQueriesData<InfiniteData<TodosResponse>>({
+  return queryClient.getQueriesData<InfiniteData<GetTodosResponse>>({
     predicate: (query) => isTodosQueryKey(query.queryKey),
   });
 };
